upload blob just after fetch data; store blobId in rollbar error and upload record
