+++ lib/core/api.js
@@ -554,6 +554,10 @@ api.upload.toPlatform = function(data, sessionInfo, progress, groupId, cb, uploa
         uploadItem.client.private.delta = sessionInfo.delta;
       }
 
+      if(sessionInfo.blobId != null) {
+        uploadItem.client.private.blobId = sessionInfo.blobId;
+      }
+
       if(uploadType === 'jellyfish') {
         uploadItem.with_uploadId('upid_' + md5(sessionInfo.deviceId + '_' + sessionInfo.start).slice(0, 12));
         uploadItem.with_guid(uuidv4());
@@ -708,7 +712,7 @@ api.errors.log = function(error, message, properties) {
   api.log('GET /errors');
 
   if (rollbar) {
-    if (error.blobId) {
+    if (error.data && error.data.blobId) {
       rollbar.error(error, { blobId: error.blobId });
     } else {
       rollbar.error(error);
