+++ lib/driverManager.js
@@ -15,6 +15,7 @@
  * == BSD2 LICENSE ==
  */
 
+var _ = require('lodash');
 var async = require('async');
 var statusManager = require('./statusManager.js');
 var debug = require('bows')('DriverManager');
@@ -116,7 +117,35 @@ module.exports = function (driverObjects, configs) {
       }
       function fetchData(data, next){
         try{
-          drvr.fetchData(stat.progressForStep(3), data, next);
+          drvr.fetchData(stat.progressForStep(3), data, function(err, recs) {
+            if(recs != null && _.isArray(recs.pages || recs.aapPackets)) {
+              /*
+                we currently support binary blobs for Medtronic (.pages) and
+                Libre (.aapPackets)
+              */
+
+              const CONTENT_TYPE = 'application/json';
+              const data = _.omit(recs, ['post_records']);
+              const filenameBinary = 'binary-blob.json';
+              const jsonDataBinary = JSON.stringify(data, undefined, 4);
+
+              configs[driver].api.upload.blob(jsonDataBinary, CONTENT_TYPE, function (err, result) {
+                if (err) {
+                  // we shouldn't fail if we can't upload the binary blob
+                  debug(err);
+                }
+
+                if (result && result.id) {
+                  debug('Blob ID:', result.id);
+                  recs.blobId = result.id;
+                }
+                console.log("RECS:", recs);
+                return next(err, recs);
+              });
+            } else {
+              return next(err, recs);
+            }
+          });
         } catch (fetchDataError) {
           fetchDataError.step = 'fetchData';
           return next(fetchDataError);
