+++ lib/drivers/abbott/abbottFreeStyleLibre.js
@@ -239,6 +239,7 @@ export default function (config) {
         timeProcessing: cfg.tzoUtil.type,
         tzName: cfg.timezone,
         version: cfg.version,
+        blobId: data.blobId,
       };
 
       cfg.api.upload.toPlatform(
