+++ lib/drivers/medtronic/medtronicDriver.js
@@ -1617,7 +1617,8 @@ module.exports = function (config) {
         start: sundial.utcDateString(),
         timeProcessing: cfg.tzoUtil.type,
         tzName : cfg.timezone,
-        version: cfg.version
+        version: cfg.version,
+        blobId: data.blobId
       };
 
       cfg.api.upload.toPlatform(data.post_records, sessionInfo, progress, cfg.groupId, function (err, result) {
